EYE_GLAZE/Python_Backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
tensorflow==2.15.0
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.1.0
scikit-learn==1.3.2
